changed how Global.L is utalised
The way Global.L was being called inside functions was not allowing it to be freely updated to adjust for the scale factor. It is now changed so that a local variable is used within the functions environment instead of contiuously updating and calling a variable from the global environment.